docs(filter-buttons): update single and multiple docs and structure
