+++ .storybook/main.js
@@ -20,7 +20,9 @@ module.exports = {
     '../projects/canopy/src/lib/forms/checkbox-group/filter-group.stories.ts',
     '../projects/canopy/src/lib/forms/date/date-field.stories.ts',
     '../projects/canopy/src/lib/forms/input/input.stories.ts',
-    '../projects/canopy/src/lib/forms/radio/filter.stories.ts',
+    '../projects/canopy/src/lib/forms/docs/filter-buttons/guide.stories.mdx',
+    '../projects/canopy/src/lib/forms/docs/filter-buttons/filter-single.stories.ts',
+    '../projects/canopy/src/lib/forms/docs/filter-buttons/filter-multiple.stories.ts',
     '../projects/canopy/src/lib/forms/radio/radio.stories.ts',
     '../projects/canopy/src/lib/forms/radio/segment.stories.ts',
     '../projects/canopy/src/lib/forms/select/select.stories.ts',
