+++ .storybook/preview-head.html
@@ -40,7 +40,8 @@
     margin: var(--space-lg) 0 var(--space-sm) !important;
   }
 
-  .sbdocs.standfirst {
+  .sbdocs.standfirst,
+  .sbdocs.standfirst .sbdocs.sbdocs-a {
     font-size: var(--text-fs-3-size) !important;
     font-weight: var(--font-weight-light) !important;
     line-height: var(--text-fs-3-line-height) !important;
